% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploysPipline.R
\name{contigency_test}
\alias{contigency_test}
\title{This function perform contigency test for a catagory or subcatagory between two cohorts}
\usage{
contigency_test(catDF, subcat = NULL)
}
\arguments{
\item{catDF}{A dataframe of catagory from combined two cohorts}
}
\value{
numeric value of p-value
}
\description{
This function perform contigency test for a catagory or subcatagory between two cohorts
}
