% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cvds}
\alias{cvds}
\title{Cardivescular disease list}
\format{An object of class \code{character} of length 5.}
\usage{
cvds
}
\description{
Cardivescular disease list
}
\examples{
\dontrun{
cvds
}
}
\keyword{datasets}
