% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploysPipline.R
\name{filehandler}
\alias{filehandler}
\title{This function combines two cohorts from Exploys}
\usage{
filehandler(file1, file2)
}
\arguments{
\item{file1}{A string of one downloaded cohort file path}

\item{file2}{A string of another downloaded cohort file path}
}
\value{
A named list and name is individual sheet name, such as "Diagnosis"
}
\description{
This function combines two cohorts from Exploys
}
