% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploysPipline.R
\name{getSigItems}
\alias{getSigItems}
\title{Perform Chi square goodness test to compute the significant catagory or sub-catagories 
with at least 100 observations}
\usage{
getSigItems(catDF, subcat = NULL, cohortDF, fdr = 0.001, min_obs = 10)
}
\arguments{
\item{catDF}{A dataframe of catagory from combined two cohorts}

\item{subcat}{A string set of sub-catagory, sucha as CVDs, default is null}

\item{cohortDF}{A dataframe containing cohort search terms}

\item{fdr}{multiple tests correcton threshold, default is 0.001}

\item{min_obs}{mininal number of observations required for comparson}
}
\value{
A dataframe with statitics test result
}
\description{
Perform Chi square goodness test to compute the significant catagory or sub-catagories 
with at least 100 observations
}
