% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploysPipline.R
\name{or_conf}
\alias{or_conf}
\title{This function computes odd ratio (OR) and confidence inteval (CI) using logistic regression}
\usage{
or_conf(sigcatDF, subcat = NULL, cohortDF, conf = FALSE)
}
\arguments{
\item{sigcatDF}{A dataframe with statistics}

\item{subcat}{A string set of sub-catagory, sucha as CVDs, default is NA}

\item{cohortDF}{A dataframe containing cohort search terms}

\item{conf}{A beelean indication if confidence inteval will be computed, defalut is false due to 
expensive computing.}
}
\value{
A dataframe with odd ratio and confidence inteval
}
\description{
This function computes odd ratio (OR) and confidence inteval (CI) using logistic regression
}
